refactor(constant): update constant file name
