+++ src/constants/graphql/POKEMON_DETAIL_QUERY.ts
