+++ src/constants/graphql/POKEMON_LIST_QUERY.ts
